requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
gunicorn==21.2.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
itsdangerous==2.2.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic[email]
argon2-cffi
cryptography
firebase-admin==6.7.0
Pillow